requirements.txt
tabulate==0.8.9
kneed==0.7.0
scipy==1.5.4
joblib==1.1.0
pandas==1.3.5
loguru==0.5.3
tqdm==4.63.1
protobuf==3.19.4
autokeras==1.0.19
pydub==0.25.1
librosa==0.7.2
numba==0.48.0
seaborn==0.11.2
matplotlib==3.5.1
imageio-ffmpeg==0.4.5
boto3==1.24.18
scikit-image==0.19.2
scikit-plot==0.3.7
pandas-profiling==3.0.0
tensorflow==2.8.2
tensorflow-datasets==4.0.1
tensorflow-estimator==2.8.0
tensorflow-gcs-config==2.8.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.8.0
tensorflow-probability==0.16.0
tensorflow-addons==0.16.1
awscli==1.25.18
psutil==5.9.1
tensorflow-decision-forests==0.2.4
ipdb==0.13.9
climage==0.1.3
xgboost==1.6.1
hyperopt==0.2.7
sqlalchemy-cockroachdb==1.4.3